harden(#11,#3,#2): DSN redaction, RAII temp-dir cleanup, 128-bit cache key
Three independent hardening fixes developed together on this branch:

#11 Credential redaction in logs: new kokedb_common::redact::redact_dsn masks
the password in any `scheme://user:pass@host` DSN; the catalog-sync log sites
now log the redacted form instead of the raw DSN (which leaked the meta-store
password into logs).

#3 RAII temp-dir cleanup: new kokedb_common::file::TempPath removes its
directory on Drop unless disarmed. The sync runner arms a guard over the
in-progress snapshot dir (disarmed only after the schema is committed) plus the
flat/delta scratch dirs, and shadow validation guards its per-run temp base —
so a failure or early return no longer leaks partial snapshots to disk.

#2 128-bit cache key: a 64-bit plan hash has a non-negligible birthday
collision probability at scale, and a collision would serve one query's cached
rows for a different query — a silent correctness bug. get_plan_hash now uses
xxh3_128; u64 -> u128 is propagated through the whole cache-key chain (query,
plan resolver, catalog manager, meta store, task manager, MySQL server, the
foyer hybrid cache, singleflight). The meta columns that persist these keys
(sql_stats.sql_hash, table_stats.cache_key_list) widen from VARCHAR(20) to
VARCHAR(40) with idempotent ALTER ... TYPE migrations for existing deployments.

+++ src/cache/src/foyer_hybrid.rs
@@ -15,7 +15,7 @@ const DEFAULT_ZSTD_LEVEL: i32 = 3;
 
 #[derive(Clone)]
 pub struct LruResultCache {
-    pub inner: Arc<HybridCache<u64, Vec<u8>>>,
+    pub inner: Arc<HybridCache<u128, Vec<u8>>>,
     compression_level: i32,
 }
 
@@ -37,7 +37,7 @@ impl LruResultCache {
                 CacheError::FoyerError(format!("Failed to build fsdevice with error: {x}"))
             })?;
 
-        let hybrid_cache: HybridCache<u64, Vec<u8>> = HybridCacheBuilder::new()
+        let hybrid_cache: HybridCache<u128, Vec<u8>> = HybridCacheBuilder::new()
             .memory(memory_size * 1024 * 1024)
             .storage()
             .with_engine_config(BlockEngineBuilder::new(device))
@@ -64,19 +64,19 @@ impl LruResultCache {
         })
     }
 
-    pub async fn insert(&self, key: u64, batches: &[RecordBatch]) -> Result<(), CacheError> {
+    pub async fn insert(&self, key: u128, batches: &[RecordBatch]) -> Result<(), CacheError> {
         let encode = RecordBatchCodec::encode(batches, self.compression_level)?;
         let _ret = self.inner.insert(key, encode);
 
         Ok(())
     }
 
-    pub async fn delete(&self, key: u64) -> Result<(), CacheError> {
+    pub async fn delete(&self, key: u128) -> Result<(), CacheError> {
         let _ret = self.inner.remove(&key);
         Ok(())
     }
 
-    pub async fn get(&self, key: u64) -> Result<SendableRecordBatchStream, CacheError> {
+    pub async fn get(&self, key: u128) -> Result<SendableRecordBatchStream, CacheError> {
         if let Some(entry) = self.inner.get(&key).await.map_err(|x| {
             CacheError::Internal(format!("Failed to get cache key: {} with error: {x}", &key))
         })? {
